[callback] Fix the embed test. Remove pyinterpreter
Replace the embed test by a callback test.
Test a callback to Python interpreter.
embedding the interpreter is a different question, which may be useful
or not. pyinterpreter is not correct as is, and not used, so remove it.

previous test was broken in gcc os x when embedding the interpreter
as scipy expects an interpreter as main.